requirements.txt
Django~=5.0.6
djangorestframework
python-decouple
psycopg2-binary
Pillow
django-cors-headers